feat: replace lazy with vimpack
