--- nvim/.config/nvim/.gitignore
@@ -1,8 +1,0 @@
-tt.*
-.tests
-doc/tags
-debug
-.repro
-foo.*
-*.log
-data
