+++ nvim/.config/nvim/init.lua
@@ -1,2 +1,6 @@
--- bootstrap lazy.nvim, LazyVim and your plugins
-require("config.lazy")
+vim.g.mapleader = " "
+
+vim.opt.rtp:prepend(vim.fn.stdpath('config'))
+require("config.options")
+require("config.keymaps")
+require("config.autocommands")
