--- nvim/.config/nvim/lazyvim.json
@@ -1,10 +1,0 @@
-{
-  "extras": [
-
-  ],
-  "install_version": 8,
-  "news": {
-    "NEWS.md": "11866"
-  },
-  "version": 8
-}
