--- nvim/.config/nvim/lua/plugins/disabled.lua
@@ -1,7 +1,0 @@
-return {
-  { "MagicDuck/grug-far.nvim", enabled = false },
-  { "folke/flash.nvim", enabled = false },
-  { "folke/which-key.nvim", enabled = false },
-  { "akinsho/bufferline.nvim", enabled = false },
-  {"nvim-lualine/lualine.nvim", enabled = false},
-}
