--- nvim/.config/nvim/lua/plugins/formatting.lua
@@ -1,28 +1,0 @@
-return {
-  {
-    "stevearc/conform.nvim",
-    event = { "BufWritePre" },
-    opts = {
-      formatters_by_ft = {
-        lua = { "stylua" },
-        rust = { "rustfmt" },
-        python = { "ruff_format" },
-        typescript = { "prettier" },
-        html = { "prettier" },
-      },
-    },
-    keys = {
-      {
-        "<leader>fo",
-        mode = { "n", "x" },
-        function()
-          require("conform").format({
-            lsp_fallback = true,
-            async = false,
-            timeout_ms = 500,
-          })
-        end,
-      },
-    },
-  },
-}
