--- nvim/.config/nvim/lua/plugins/harpoon.lua
@@ -1,20 +1,0 @@
-return {
-  "ThePrimeagen/harpoon",
-  opts = {},
-  keys = {
-    {
-      "<leader>h",
-      function()
-        require("harpoon.ui").toggle_quick_menu()
-      end,
-      desc = "toggle harpoon ui",
-    },
-    {
-      "<leader>ha",
-      function()
-        require("harpoon.mark").add_file()
-      end,
-      desc = "toggle harpoon ui",
-    },
-  },
-}
