--- nvim/.config/nvim/lua/plugins/lang/helm.lua
@@ -1,4 +1,0 @@
-return {
-  -- https://github.com/mrjosh/helm-ls?tab=readme-ov-file#editor-config-examples
-  { "qvalentin/helm-ls.nvim", ft = "helm" },
-}
