--- nvim/.config/nvim/lua/plugins/lang/python.lua
@@ -1,17 +1,0 @@
-return {
-  "neovim/nvim-lspconfig",
-  opts = {
-    servers = {
-      ty = {
-        settings = {
-          ty = {
-            diagnosticMode = "workspace",
-            experimental = {
-              rename = true,
-            },
-          },
-        },
-      },
-    },
-  },
-}
