--- nvim/.config/nvim/lua/plugins/lang/rust.lua
@@ -1,92 +1,0 @@
-return {
-  -- lsp-config
-  {
-    "neovim/nvim-lspconfig",
-    opts = {
-      -- this globally disables stuff
-      inlay_hints = { enabled = false },
-      servers = {
-        rust_analyzer = {
-          filetypes = { "rust" },
-          settings = {
-            ["rust-analyzer"] = {
-              checkOnSave = true,
-              diagnostics = {
-                enable = true,
-              },
-              cargo = {
-                features = "all",
-                buildScripts = {
-                  enable = true,
-                },
-              },
-              procMacro = {
-                enable = true,
-              },
-              files = {
-                excludeDirs = {
-                  ".direnv",
-                  ".git",
-                  ".github",
-                  ".gitlab",
-                  "bin",
-                  "node_modules",
-                  "target",
-                  "venv",
-                  ".venv",
-                },
-              },
-            },
-          },
-        },
-      },
-    },
-  },
-  -- crates.nvim
-  {
-    "saecki/crates.nvim",
-    tag = "stable",
-    event = { "BufRead Cargo.toml" },
-    opts = {
-      completion = {
-        crates = {
-          enabled = true,
-        },
-      },
-      lsp = {
-        enabled = true,
-        actions = true,
-        completion = true,
-        hover = true,
-      },
-    },
-  },
-  -- rustacean
-  -- {
-  --   "mrcjkb/rustaceanvim",
-  --   opts = {
-  --     server = {
-  --       default_settings = {
-  --         ["rust-analyzer"] = {
-  --           checkOnSave = false, -- disable for large projects
-  --           diagnostics = {
-  --             enable = true,
-  --           },
-  --           cargo = {
-  --             buildScripts = {
-  --               enable = true,
-  --             },
-  --           },
-  --           procMacro = {
-  --             enable = true,
-  --             ignored = {}, -- should allow for async-trait :(
-  --           },
-  --         },
-  --       },
-  --     },
-  --   },
-  --   config = function(_, opts)
-  --     vim.g.rustaceanvim = vim.tbl_deep_extend("keep", vim.g.rustaceanvim or {}, opts or {})
-  --   end,
-  -- },
-}
