--- nvim/.config/nvim/lua/plugins/linting.lua
@@ -1,11 +1,0 @@
-return {
-  "mfussenegger/nvim-lint",
-  event = "LazyFile",
-  opts = {
-    -- Event to trigger linters
-    events = { "BufWritePost", "BufReadPost", "InsertLeave" },
-    linters_by_ft = {
-      -- typescript = { 'eslint' },
-    },
-  },
-}
