--- nvim/.config/nvim/lua/plugins/treesitter.lua
@@ -1,18 +1,0 @@
-return {
-  "nvim-treesitter/nvim-treesitter",
-  opts = {
-    ensure_installed = {
-      "rust",
-      "ron", -- used with serde ?
-      "python",
-      "typescript",
-      "hurl",
-      "yaml",
-      "terraform",
-      "hcl",
-      "helm",
-      "c", -- included by default
-      "go",
-    },
-  },
-}
