+++ nvim/.config/nvim/plugin/lang/helm.lua
@@ -1,0 +1,3 @@
+vim.pack.add({
+	"https://github.com/qvalentin/helm-ls.nvim",
+})
