+++ nvim/.config/nvim/plugin/lang/python.lua
@@ -1,0 +1,10 @@
+vim.lsp.config("ty", {
+	settings = {
+		ty = {
+			diagnosticMode = "workspace",
+			experimental = {
+				rename = true,
+			},
+		},
+	},
+})
