--- nvim/.config/nvim/stylua.toml
@@ -1,3 +1,0 @@
-indent_type = "Spaces"
-indent_width = 2
-column_width = 120
